Solution of all given problem . 1
Write a program in Java to create a thread by extending the ‘Thread’
class and by implementing the “Runnable” interface

2

Write a program in Java to demonstrate sleep() and wait()

3

Write a program in Java to demonstrate synchronization

4

Write a program in Java to demonstrate try and catch

5

Writing a program in Java to throws, throw, finally, and custom
exceptions in Java.

6

Write a program in Java to demonstrate exception handling

7

Writing a program in Java to create, read, update, and delete operations
on the files in Java.

8

Write a program in Java to demonstrate the uses of classes, objects, and
the object-oriented pillars in Java

9

Write a program in Java to resolve the diamond problem using OOPs’
concepts